backend/pyproject.toml
[project]
name = "flexist-backend"
version = "0.1.0"
description = "Flexist - Gym Management & Fitness Tracking API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]",
    "sqlmodel>=0.0.22",
    "psycopg2-binary",
    "alembic",
    "python-jose[cryptography]",
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "pydantic-settings",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
